flux/ai/agent-sre/src/requirements.txt
# === Core Framework ===
fastapi>=0.115.6
uvicorn>=0.32.1
pydantic>=2.12.5
python-dotenv>=1.0.0
cloudevents>=1.10.0

# === MLX-LM Framework (for FunctionGemma) ===
# Note: MLX requires Apple Silicon or compatible hardware
# For non-Apple Silicon, use alternative inference backend
mlx>=0.19.0
mlx-lm>=0.20.0

# === LLM Integration ===
# Ollama for local inference (alternative to MLX)
ollama>=0.3.0
anthropic>=0.30.0  # Fallback to Claude
httpx>=0.27.0

# === Prometheus Integration ===
prometheus-client>=0.21.0
prometheus-api-client>=0.5.3

# === Langchain & DeepAgents (Evaluation) ===
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
# deepagents - install from GitHub if compatible
# git+https://github.com/langchain-ai/deepagents.git

# === LangGraph for Workflow Orchestration ===
langgraph>=0.2.0
langgraph-checkpoint>=0.2.0

# === Agent-Lightning for Optimization ===
# Note: agent-lightning may need to be installed from source or specific version
# agent-lightning>=0.1.0  # Uncomment when available

# === Observability ===
opentelemetry-api>=1.30.0
opentelemetry-sdk>=1.30.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-httpx>=0.45b0
opentelemetry-exporter-otlp-proto-grpc>=1.30.0
opentelemetry-exporter-otlp-proto-http>=1.30.0
prometheus-client>=0.21.0
structlog>=24.4.0

# === Utilities ===
tenacity>=8.3.0
pyyaml>=6.0.0
jinja2>=3.1.0
python-dateutil>=2.9.0

# === Data Processing ===
pandas>=2.2.0
numpy>=1.26.0

# === RAG (Retrieval Augmented Generation) ===
# Optional: sentence-transformers for semantic embeddings
# sentence-transformers>=2.2.0  # Uncomment if using RAG with embeddings

# === Report Generation ===
markdown>=3.7.0
jinja2>=3.1.0

# === Testing ===
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

